moves CachedPeerAddrResponse to a module
